a basic version to make block log precess series service written easier
